// lib/schemas/create-billing-schema.ts
export const BillingProviderSchema = z.enum([
  'stripe',
  'paddle',
  'lemon-squeezy',
]);
// ...
export const ProductSchema = z
  .object({
    id: z
      .string()
      .describe('Unique identifier for the product. Defined by th Provider.')
      .min(1),
    name: z
      .string()
      .describe('Name of the product. Displayed to the user.')
      .min(1),
    description: z
      .string()
      .describe('Description of the product. Displayed to the user.')
      .min(1),
    currency: z
      .string()
      .describe('Currency code for the product. Displayed to the user.')
      .min(3)
      .max(3),
    badge: z
      .string()
      .describe(
        'Badge for the product. Displayed to the user. Example: "Popular"',
      )
      .optional(),
    features: z
      .array(
        z.string().describe('Features of the product. Displayed to the user.'),
      )
      .nonempty(),
    enableDiscountField: z
      .boolean()
      .describe('Enable discount field for the product in the checkout.')
      .optional(),
    highlighted: z
      .boolean()
      .describe('Highlight this product. Displayed to the user.')
      .optional(),
    plans: z.array(PlanSchema),
  })
  .refine((data) => data.plans.length > 0, {
    message: 'Products must have at least one plan',
    path: ['plans'],
  })
  .refine(
    (item) => {
      const planIds = item.plans.map((plan) => plan.id);

      return planIds.length === new Set(planIds).size;
    },
    {
      message: 'Plan IDs must be unique',
      path: ['plans'],
    },
  );

const BillingSchema = z
  .object({
    provider: BillingProviderSchema,
    products: z.array(ProductSchema).nonempty(),
  })
  .refine(
    (data) => {
      const ids = data.products.flatMap((product) =>
        product.plans.flatMap((plan) => plan.lineItems.map((item) => item.id)),
      );

      return ids.length === new Set(ids).size;
    },
    {
      message: 'Line item IDs must be unique',
      path: ['products'],
    },
  )
  .refine(
    (schema) => {
      if (schema.provider === 'lemon-squeezy') {
        for (const product of schema.products) {
          for (const plan of product.plans) {
            if (plan.lineItems.length > 1) {
              return false;
            }
          }
        }
      }

      return true;
    },
    {
      message: 'Lemon Squeezy only supports one line item per plan',
      path: ['provider', 'products'],
    },
  )
  .refine(
    (schema) => {
      if (schema.provider !== 'lemon-squeezy') {
        // Check if there are any flat fee metered items
        const setupFeeItems = schema.products.flatMap((product) =>
          product.plans.flatMap((plan) =>
            plan.lineItems.filter((item) => item.setupFee),
          ),
        );

        // If there are any flat fee metered items, return an error
        if (setupFeeItems.length > 0) {
          return false;
        }
      }

      return true;
    },
    {
      message:
        'Setup fee metered items are only supported by Lemon Squeezy. For Stripe and Paddle, please use a separate line item for the setup fee.',
      path: ['products', 'plans', 'lineItems'],
    },
  );

export function createBillingSchema(config: z.infer<typeof BillingSchema>) {
  return BillingSchema.parse(config);
}

export type BillingConfig = z.infer<typeof BillingSchema>;
export type ProductSchema = z.infer<typeof ProductSchema>;